make non-postgres default, update script, test
