+++ .python-version
@@ -1,1 +1,1 @@
-3.12
+3.13
