+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "dr-wandb"
-version = "0.1.1"
+version = "0.1.2"
 description = "Interact with wandb from python"
 readme = "README.md"
 authors = [
@@ -9,15 +9,14 @@ authors = [
 requires-python = ">=3.12"
 dependencies = [
     "pandas>=2.3.2",
-    "psycopg2>=2.9.10",
     "pyarrow>=21.0.0",
-    "pydantic-settings>=2.10.1",
-    "sqlalchemy>=2.0.43",
+    "typer>=0.20.0",
     "wandb>=0.21.4",
+    "sqlalchemy>=2.0.43",
 ]
 
 [project.scripts]
-wandb-download = "dr_wandb.cli.download:download_project"
+wandb-download = "dr_wandb.cli.download:app"
 
 [build-system]
 requires = ["hatchling"]
@@ -25,6 +24,13 @@ build-backend = "hatchling.build"
 
 [tool.hatch.metadata]
 allow-direct-references = true
+
+[tool.hatch.build]
+exclude = [
+  "docs",
+  "docs/**",
+]
+
 
 [tool.hatch.build.targets.wheel]
 packages = ["src/dr_wandb"]
@@ -34,6 +40,9 @@ packages = ["src/dr_wandb"]
 dev = [
     "dr-wandb",
     "pytest>=8.4.1",
+]
+postgres = [
+    "pydantic-settings>=2.10.1",
 ]
 
 [tool.ruff]
