+++ src/dr_wandb/constants.py
@@ -1,6 +1,7 @@
 from collections.abc import Callable
 from typing import Literal
 
+from sqlalchemy import String
 from sqlalchemy.orm import DeclarativeBase
 
 
@@ -16,5 +17,7 @@ type FilterField = Literal["project", "entity", "state", "run_ids"]
 WANDB_RUN_STATES = ["finished", "running", "crashed", "failed", "killed"]
 type RunState = Literal["finished", "running", "crashed", "failed", "killed"]
 type RunId = str
+
+Base.type_annotation_map = {RunId: String}
 
 type ProgressCallback = Callable[[int, int, str], None]
